Fix to camera stuck bug (#10132)
## Describe your changes
When we added the camera switch functionality, we added this `useEffect`
```
useEffect(() => {
    if (mediaStream && videoRef.current ) {
      videoRef.current.srcObject = mediaStream;
    }
  }, [mediaStream]);
```

I believe this `useEffect` was colliding with:

```
if (mediaStream && videoRef.current && !videoRef.current.srcObject) {
    videoRef.current.srcObject = mediaStream;
  }
```

So, I introduced a new variable in this PR `isInitialMediaStream` so
that they don't collide. Also, we don't want to switch until the video
starts playing, and restart playing after the switch. If the switching
PR created the bug, this PR makes it so that the behavior/logic before
switching happens (i.e. when `isInitialMediaStream == true`) is exactly
as it was before the switching PR.


## Caveats


## Test Plan