Add API to get VDR AWS pre-enrollment status (#10136)
## Describe your changes
Will allow the CLI to give guidance that the external ID needs to be generated for use in the IAM role before continuing enrollment.

I think it's useful to keep pre-enrollment (external ID generation) and enrollment separate command, since there is work the tenant has to do in between that may take some time. This API allows us to give a hint that pre-enrollment hasn't been started.

## Caveats

## Test Plan