Implement enrollment keygen for Vault DR (#10208)
## Describe your changes
Most of the diff line count is `Cargo.lock` and a moved file.

See the [How it Works](https://www.notion.so/onefootprint/Vault-Disaster-Recovery-Docs-Draft-8edb880bb1af4576b6bbfbb71e547ad5) section of the docs.

The client generates the org key pair, passes the public key to the server. Then, the server generates the recovery key pair, wraps the org recovery private key with the org public key, and immediately discards the recovery private. The recovery public key is stored in the DB so we can files to that recipient. The wrapped recovery key (the recovery payload, only usable by the tenant), is stored in the DB so we can retrieve it.

Maybe it would make sense to store the recovery payload encrypted with KMS. Then, we can ensure that only an account admin can decrypt, and the app can only encrypt. This is straight-forward modification of this starting point if we decide to do this.

## Caveats

## Test Plan