Produce request for samba from OCR response instead of loading from vault (#10225)
## Describe your changes
The goal of samba is to check information off the DL to the DMV. We want
to optionality to run this adhoc for:
* an fp_id
* for data
* for a specific DL collected

As part of this, we need to construct a view of the data we got off the
DL and send it to samba in the format they want.

One way to do this is to fetch versioned DIs from the vault. This would
support fp_id and specific DL use cases above. however,
- not as straightfoward as using an identifier like doc_id
- anything we want to send like first name/last name (as we only vault
full_name currently) or something like height/weight/eye color etc now
requires vaulting.


The other way is just to retrieve the OCR response associate with a
doc_id and build a request to samba.

Other notes
- In the current implementation, we fetch IDKs which even in doc first
isn't smart. Users can change things.
- On one view however, it's kinda a smart fraud solution "does the
IDK::Name and the DL number go together"? however, this is confusing.
it's simpler and more straightforward to compute
   - Does DL info match DMV &&
   - Does IDK::Name match OCR name (which we do already)

## Caveats


## Test Plan