Add table to track progress writing Vault DR backups (#10212)
## Describe your changes
~The API will create an entry in the `vault_dr_write_task` table when we write a new data lifetime. The background Vault DR worker will read incomplete tasks and write `vault_dr_record` rows, and update the `vault_dr_write_task` `completed_at` timestamp.~

Changed to this:
An alternative architecture could be to make tasks implicit based on the current state of the `data_lifetime` table. That may make more sense if there's not a good central place to hook into writing `data_lifetime`s.

Opted for the "implicit task" alternative to avoid reading Vault DR config in VaultWrapper and adding complexity to that central component.

## Caveats

## Test Plan