Only send onboarding.status_changed webhook for terminal statuses (#10265)
## Describe your changes
Will need this in order to make the refactor in the PR above this

This does make a breaking change to how we send the deprecated
`onboarding.status_changed` webhook: before this PR, we would send it
for silly status changes between pending / incomplete. Now, we only send
this webhook when decisions are made (either manual review or footprint
decisions). I cannot imagine any tenants are paying attention to the
status updates between pending <> incomplete... we can keep the old
behavior, but it just makes the code a little uglier. Let me know if you
think worth keeping



## Caveats


## Test Plan