Fix requesting additional info (#10328)
## Describe your changes
when a tenant requests a user to redo document, we create a document
workflow. all workflows require an obc_id in order to function. doc
workflows aren't associated with a playbook directly (since they are out
of band). instead, they are associated with the latest "reonboardable"
playbook

so if a tenant requests to redo doc, if the latest playbook has been
deactivated, when a user goes to start the workflow, we check OBC.status
and error

changing workflow to not need an obc_id is involved, so the simplest fix
is just to error in the trigger request and point a tenant to creating a
Document only playbook or a new playbook.

## Caveats
has the side effect of making it more clear for incomplete users too

## Test Plan
itest