[BE-306] Clean up dashboard implementation to decrypt documents (#10361)
## Describe your changes
We have this very general `DecryptForm` wrapper around all of the vault
data section on the dashboard. The DecryptForm passes which fields are
selected to be decrypted to our decrypt state machine, which keeps track
of the DataIdentifier fields to request from the backend API.

Before this PR, we tried to squeeze documents into this DataIdentifier
paradigm - we assigned each document type its own "main" DI, and the
decrypt machine would understand that if a given DI is provided, we
should decrypt the document type represented by that main DI.

Now, we are much more explicit with our intent - we pass the document
type itself to the decrypt machine, and the decrypt machine maps the
document type to the list of DIs that need to be decrypted.

Similarly, we were trying to squeeze documents into the same paradigm as
other DIs by using the `<Field />` component and `useField` hook to
render documents. Documents are and will continue to be sufficiently
different from the single-DI model of the rest of the vault, so I made
specialized components to render documents

## Caveats
In our [future document
design](https://www.figma.com/design/5VkT8cIz0fJQsQFGaa5iN6/%F0%9F%91%A8%F0%9F%8F%BE%E2%80%8D%F0%9F%A6%B3-Users?node-id=7342-118844&t=MPDxWUcoNrnSZPW5-4),
we every document that the backend knows about, rather than our current
designs where we group them by document type. In this world, we'll
probably want to pass the document IDs (literal unique identifiers) into
the decrypt machine. This at least starts to blaze a trail to do so!


## Test Plan


https://github.com/onefootprint/monorepo/assets/21066341/ea270ea3-e66f-4db2-818a-67216c50d63a

NOTE: there's a bug with custom docs shown in the video above when doing
"decrypt all" - i've fixed it in this PR


And another test case for a user who doesn't have access to decrypt
documents:


https://github.com/onefootprint/monorepo/assets/21066341/37c63bf5-0b1d-4261-88b7-d80524cb261d