Show correct classification of documents in dashboard (#10363)
## Describe your changes
After the refactor in the PR below this, the frontend is no longer
inferring which _image_ DIs to decrypt based on the document kind we
return here, so we're able to change the type of the document we return
here.

The real benefit of returning vaulted_document_type here is that the
frontend will now properly infer the correct OCR DIs to decrypt. After
this change, we now will show OCR data and images properly for an image
that was misclassified by the user!

## Caveats
The frontend is still hardcoding which OCR DIs to decrypt based on the
document kind returned here. We could change this in the future.
## Test Plan