Business Insights BE API v0 pt1/2 (#10355)
## Describe your changes
Creates a function that transforms a middesk business response into a format that can be used for the designs for the business insights page: https://www.figma.com/design/ZEfaxUtluv2wA1zytKyAZP/%F0%9F%8F%A2-Businesses?node-id=244-30396&t=oHHRfoA8twgyyfC0-0

There's a couple of small details to work out, but ~96% of things are there

## Caveats

## Test Plan
will add unit tests for specific aspects, and itests in the API route integration. I've ran a bunch of real middesk responses through this and printed out the json it generates and compared to the middesk dashboard/api response and it seems to be doing the right thing. but there's edge cases I just don't know how middesk handles. the real tests are just going to be rolling this out behind a flag and checking a lot of examples i think