[1/2][BE-333] Explicitly save the business workflow ID in auth session (#10437)
## Describe your changes
Now that we have multiple avenues for making a business workflow during
bifrost, it helps to track explicitly when a business workflow has been
created for an auth session. So we now store it in the encrypted auth
session, just as we already store the person `workflow_id`.

This starts to write it to the auth session. On the read path, we will
read _either_ the written `biz_wf_id` or fall back to looking it up via
the `sb_id` for backcompat. In a future PR, we'll migrate to only read
from the new path

## Caveats


## Test Plan