Parse out database connection closed error (#10534)
## Describe your changes
Lots of digging to find this little line 😛 

When the DB segfaulted at the offsite, the DB instance restarted within
~20 seconds, but we experienced a downtime of ~2 minutes. We kept trying
to reuse connections that were killed forcefully by the server. So, this
is the start of being able to parse when there's an error from which the
client cannot recover. Eventually, we'll start responding to these
errors by removing the object from the pool.

I found this from this PR in diesel that migrates to successfully parse
when the connection is closed by the server:
https://github.com/diesel-rs/diesel/pull/3135

## Caveats


## Test Plan
Locally, made a request:
```
$ curl -s localhost:8000/users -u sk_test_bO5lzm9r2HLlloOBeizIqmPM3tILO94n1i: | jq
```
This opened up a connection and saved it inside the DB pool. This was
the only connection in the pool.

Then, in a pg shell, killed the connection.
```
footprint_db=# select pid, query from pg_stat_activity;
  pid  |                                                                                            query
-------+----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
...
 16342 | SELECT COUNT(*) FROM "scoped_vault" WHERE (((("scoped_vault"."tenant_id" = $1) AND ("scoped_vault"."is_live" = $2)) AND ("scoped_vault"."is_active" = $3)) AND ("scoped_vault"."kind" = $4))

footprint_db=# select pg_terminate_backend(16342);
 pg_terminate_backend
----------------------
 t
(1 row)
```

Now, the database pool still has the connection, but it has been
forcefully closed by the server. So any database queries using this
connection will continue to fail
```
$ curl -s localhost:8000/users -u sk_test_bO5lzm9r2HLlloOBeizIqmPM3tILO94n1i: | jq
{
  "message": "The DB connection has been closed",
  "code": null,
  "support_id": "8c041cd4-1c49-4aa6-91a0-f09c8854ec72",
  "debug": "ConnectionClosed"
}
```

Not relevant here, but interestingly, subsequent attempts to use the
connection fail with a different error:
```
$ curl -s localhost:8000/users -u sk_test_bO5lzm9r2HLlloOBeizIqmPM3tILO94n1i: | jq
{
  "message": "There was a problem talking to the DB: UnableToSendCommand. message: no connection to the server\n",
  "code": null,
  "support_id": "ee9b36e2-ac1e-4982-92c1-8590f21fa291",
  "debug": "OtherDieselDbError(UnableToSendCommand, \"message: no connection to the server\\n\")"
}
```