Fix samba response types (#10542)
## Describe your changes
ScrubbedPiiJsonValue cannot be used as a type on an API response as it
does _not_ actually do the scrubbing. it should only actually be used as
a type on VerificationResult - e.g. it should be a type that represents
a deserialized vendor response that _has been scrubbed_

I'll follow up and refactor so that this is clear

## Caveats


## Test Plan