Update `wal_level`, create postgres 16 parameter group (#10567)
## Describe your changes
`rds.logical_replication = 1` ends up setting the postgres setting
`wal_level` to `logical`. this is needed in order to conduct blue/green
deployments (and generally to use logical replication). my understanding
is the downside is this uses a bit more disk space. but the previous DBs
i've run all had `wal_level` set to logical by default.

this also makes a duplicate postgres 16 parameter group. i'm starting to
play around with blue/green deployments in dev, and it will be nice to
have the parameter group that we'll use pre-created in the console

## Caveats
`rds.logical_replication` requires a reboot to be applied. i tested in
dev that updating the parameter group does not force reboot the
instance.

we'll have to do a minor version upgrade from 14.5 -> 14.9 in order to
use AWS's blue/green deployments, which will require a restart. the hope
is that we can update the parameter group now and then when we restart
for the upgrade, it will be applied

## Test Plan