Fix ScrubbedPiiJsonValue and respone (#10551)
## Describe your changes
We store `response` on verification result which has the sensitive
fields scrubbed. Scrubbing is achieved via serde attributes on
individual fields


Currently we have a a struct called `ScrubbedPiiJsonValue` it is used to
represent "a deserialized vendor response has been re-serialized and
scrubbed". This is what's passed into `VerificationResult::create`.

However, in some places (mainly my fault), this struct has crept in to
being used for individual fields. Since it doesn't have any serde
attributes, it does not scrub. This PR starts a refactor to accomplish
what we actually want:
* serialization side ability to scrub individual fields from getting
stored in PG
* representing a scrubbed response

## Caveats
we could potentially remove response from verification result. however,
it's very useful to query on non-pii fields in PG, so i'm hesitant to
remove

## Test Plan
wrote test