Implement `footprint-dr list-records` (#10617)
## Describe your changes
This provides output like the following:
```
{"fp_id": "fp_id_rwSYNkPIrDFogXYjCJ1KrQ", "fields": ["id.first_name", "id.last_name", "id.phone_number"]}
{"fp_id": "fp_id_4acxG4NxFTlGaE6b2WyDIc", "fields": ["id.first_name", "id.last_name", "id.phone_number"]}
{"fp_id": "fp_id_WFVkUyTeuWNy0GlrOPdPBO", "fields": ["id.first_name", "id.last_name", "id.phone_number"]}
```

We can probably make this the last thing we implement before we ship to Grid. This allows them to make sure their backups have the right structure, at least. We can work on the decryption client while they onboard -- this is the last piece, other than the manifest correctness work which is less urgent.

## Caveats

## Test Plan