[BE-382] Remove async email verification email (#10698)
## Describe your changes

https://onefootprint.slack.com/archives/C06PWUC99NJ/p1721249461856599?thread_ts=1721242111.201349&cid=C06PWUC99NJ

I also realized through going through this process two things that
support us removing it:
- There are already a few tenants who have explicitly requested us to
stop sending these emails to their users
- We have a bug where we basically won't send this for most existing
users. Since we started putting the email into the vault during the
singup challenge when the user is created, we no longer hit this old
codepath in `PATCH /hosted/user/vault`. So we've effectively already had
this codepath removed... it's [only been sent 200 times in the last 2
months](https://ui.honeycomb.io/footprint-2e/environments/prod/datasets/fpc-api/result/56JbVqMcwS8?hideCompare)

## Caveats
There's a lot more code we can remove if we truly never want to support
this - for ex, the frontend app that we use just to hit the backend API
to mark an email as verified, the email verification session token, the
API to verify the email, etc.

We can simmer with this gone for a while and then decide if we want to
do the full cleanup

## Test Plan