Remove HydratedVendorResults (#10703)
## Describe your changes
more risk code cleanup, migrating to `loaders`

all of this will lead to us removing `VendorResult` which we have around so we can have a common type that all vendor responses can interact with. however, there's only 1 place in code currently that needs this which is the waterfall. instead we can utilize trait objects

https://linear.app/footprint/issue/BE-364/removefix-vendorresult

## Caveats

## Test Plan