Add document content type to VDR backups using S3 object metadata (#10716)
## Describe your changes
The footprint-dr client needs to be aware of a) whether a blob
represents a file instead of a regular vault field and b) what the
content type of that file is so it can add an appropriate file
extension. The easiest way to do this is to set unencrypted metadata
fields on VDR blobs that include this mime type. While I'm modifying
this behavior, I marked the proper age content type and also set the
content length on the upload request, which hopefully gives the S3
client more information to make smart handling decisions.

## Caveats


## Test Plan