[BE-398][docs] Add basic login page (#10738)
## Describe your changes
The docs site will eventually operate in a logged-in context, showing tenant-specific documentation. I'm going to build this by having the dashboard redirect to the docs site with an auth token properly scoped for docs.

This builds the docs landing page that will receive the auth token in the URL fragment and save it in session local storage.

## Caveats
I can't figure out why, but it looks like we're sending two `GET /org/member` requests to the backend upon login. LMK if anything sticks out to you :)
EDIT: ohhh maybe this is that strict mode setting that is disabled in dev - i kind of remember something about useEffect hooks being called twice locally

## Test Plan
Tested locally:

https://github.com/user-attachments/assets/c184980d-97bf-4e85-91b7-c78d16a109fc