[docs] Add logged-in user info to sidebar footer (#10740)
## Describe your changes
First part of the docs site to start showing logged-in context! We'll soon start to show docs specific to the logged-in tenant

## Caveats
This doesn't yet implement the rest of the design changes to the footer on the docs site (figma [here](https://www.figma.com/design/wu4sZKAbsxc9c8wlSXpq03/%F0%9F%93%9D-%5BDesktop%5D-Documentation?node-id=1529-86253&m=dev)). It looks a little funny right now when there is a logged in user, but no real customers will be able to get to the logged-in view until we build support for it :)

## Test Plan

https://github.com/user-attachments/assets/af602540-7a5c-4f0a-9130-6c14bf728de7