Create a new auth extractor for preview-gated APIs (#10745)
## Describe your changes
To show info on the docs site on which preview APIs are visible to the tenant, we need to actually serialize some information on what preview API guards are required for each API in the open API spec.

This adds a fun new `TenantApiKeyGated<T>` struct - it's an auth extractor, and the `T` is a type representing the preview API (which we've autogenerated from the PreviewApi enum using the `EnumVariantType` derive macro). For the `TenantApiKeyGated<T>` to be properly extracted, it actually checks the preview API guard inline! So we have a compiler guarantee that the HTTP handler function has access to the preview API.

Using `TenantApiKeyGated<T>`, we also have a custom `Apiv2Schema` implementation that serializes in the spec which preview API guard is needed to access the API.

As an aside: we could actually do the same with our normal tenant IAM role checking. This would let us serialize what kind of permissions are needed by the API key's role in order to access the API, all in the docs site!

## Caveats

## Test Plan
Look at the generated open API specs for these APIs