[1/2] Deprecate WorkosAuth session kinds (#10799)
## Describe your changes
I'm working on some of this login code on the frontend, and there's
simply just a lot of code needed to select the tenant to log into when
the user is part of multiple orgs.

No reason we can't just select the org that the user most recently
logged into - once they are logged in, they can still easily switch to
other tenants in the dropdown menu

## Caveats
The one piece of information that the "organizations" selector screen
showed was for a very-little used feature that allows generating a
dashboard link that signs you into a specific tenant. This was only used
from a coast demo flow - I'm not sure if it's still in any use. In any
case, I migrated this error message to a toast in the follow-up PR

## Test Plan