Remove hard dependency on FP API for VDR S3 metadata queries (#10820)
## Describe your changes
In order to use the footprint-dr tools without the Footprint API, we need to cut the hard dependency on the status API for fetching the bucket/namespace. The changes in this PR allow users to pass those parameters explicitly. We test with a bogus API address to ensure the API isn't used.

## Caveats

## Test Plan