Fix serialization of `DocumentFixtureResult` (#10876)
## Describe your changes
The evil `derive_more::Display` strikes again....

It is not always safe to just swap out the `Display` implementation. I
did an audit of the uses of `DocumentFixtureResult` and it does look
like it's safe to swap - was only being used in the recent inclusion in
SdkArgs in #10852. Part of what makes this safe is the util to serialize
this as a diesel Text column uses the `AsRef` implementation rather than
the `Display` implementation of the enum

## Caveats


## Test Plan