Add support for sandbox outcome in web OBC and sdk-args (#10878)
## Describe your changes

The issue was raised when @clodoan needed to skip the sandbox outcome
page in the onboarding components flow. So, we thought it'd be good to
support sandbox outcome to be passed using our sdks. This PR supports
sandbox outcome to be passed in `footprint.init` and in react onboarding
components.


## Caveats


## Test Plan