Store entire document requirement in document state machines (#10958)
## Describe your changes
The requirements are static objects, so they're safe to pass around
between the state machines. This gives a few big benefits:
- You can extremely efficiently check where a field from the onboarding
requirement is used using "Find all references." You no longer need to
track down all the codepaths where a field is destructured and passed
into other state machines
- Similarly, it is super easy to update the onboarding requirement's
API, which will inevitably happen as the product evolves. Only update
the requirement's typescript type in one place, and its context is
magically propagated into all of the state machines that use that
requirement

I think we should generally take this approach in all bifrost code we
write to handle requirements - the APIs for onboarding requirements will
inevitably change, and it's really nice to have only a few places to
update and audit when making changes to the API

## Caveats


## Test Plan