[docs] Use generic markdown component in API reference descriptions (#11012)
## Describe your changes
I previously hand wrote this markdown parsing logic, not sure why i
didn't just notice that we could reuse our markdown rendererer 😂

This is also great because we can use much more advanced formatting
logic in our descriptions

Here's the new look on the left and old look on the right. By default,
it used this more primary font color, which I kind of prefer to the old
tertiary font color. In our newer docs site, it looks like we will be
using the primary color for descriptions too


![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xncBfoFfDL8XmXxkoXEI/105894c1-f231-4fa7-ac44-ac9a112b3297.png)

## Caveats


## Test Plan