First draft of handwritten API docs (#11028)
## Describe your changes

Very basic proof of concept skeleton allows writing a markdown file for
a section of APIs. The markdown file also specifies the list of APIs
that should be included in the section. We join the section description
in the markdown file with the open API specs of the listed APIS that
should be included in the section

<img width="1394" alt="image"
src="https://github.com/user-attachments/assets/544b3171-2fb4-4902-b69b-cf85c8e579f6">

Adds a new route that we don't link to anywhere, so no tenants should be
finding this site

## Caveats


## Test Plan