Build enclave image with Packer, add Datadog agent (#11015)
## Describe your changes

Instead of doing all setup on startup (hard to debug, slow, not
deterministic), we use Pulumi with Ansible to create pre-baked AMIs for
the enclave. At runtime, we use cloud-init to template in secrets (as we
did before).

To build an image for now, just log in to the image-cache account, run
`make`, and paste in the AMI ID that's printed into the Pulumi config.
This takes about 6 minutes.

This also allows us to reuse config across machine definitions more
easily (e.g. jumpboxes, nitro enclave-compatible CI runners, nitro dev
boxes, etc.).

There are a few changes in functionality here:
- I'm upgrading to Amazon Linux 2023, from Amazon Linux 2
- Adding the DD agent to enclave nodes for metrics, logs, and (soon) APM
- Some cleanup around systemd config -- e.g. making the tailscale
connect script a oneshot so it loop until it's rate limited, running the
enclave downloads as systemd oneshots, etc.

## Caveats


## Test Plan