[BE-483] enable stepup for all sandbox configs (#10976)
## Describe your changes
In a small movement towards better demoing, we actually support a
sandbox outcome for stepup currently, but it's only allowed if it's an
alpaca CIP.

Instead, this PR changes it to dynamically allow the sandbox stepup
outcome based on whether or not there's a stepup rule configured in the
playbook. This isn't perfect since we will always do an identity doc
stepup regardless of what is chosen, but I think for demo purposes it's
slightly more in line with what we want?

The other weird aspect is writing a stepup rule then choosing in the
outcome to not do a stepup, but ultimately i think that's probably fine
to have better functionality in demos. open to thoughts!

the other idea is to just always allow stepups as a potential sandbox
outcome, though i think tying it to rules is cooler

## Caveats


## Test Plan
wrote a itest