chore(kyb) apply datadog action names (#11176)
## Describe your changes

This PR prevents PII from label, placeholder, aria-label get exposed on
DataDog RUM


![image](https://github.com/user-attachments/assets/51cae52d-dd43-4f4d-86c4-c8073ac12492)



### [How action names are
computed](https://docs.datadoghq.com/real_user_monitoring/browser/tracking_user_actions/#how-action-names-are-computed)
The Datadog Browser SDK uses different strategies to compute click
action names:

If the `data-dd-action-name` attribute or a custom attribute (as
explained above) is explicitly set by the user on the clicked element
(or one of its parents), its value is used as the action name.

If `data-dd-action-name` attribute or its equivalent **is not** set,
depending on the element type, the sdk uses other attributes such as
label, placeholder, aria-label from the element or its parents to
construct the action name. If none of these attributes is found, the sdk
uses the inner text as name for the action.