[BE-511] Add request/response fields to cmd+k (#11173)
## Describe your changes
A little more useful than some of the parameters. I'm still not so sure this provides a lot of utility... could go either way. Stripe's is realllly good at serving you relevant results for what you're searching for, but i think it will be really hard for us to build something as good as stripe's 😛

<div class='graphite__hidden'>
          <div>🎥 Video uploaded on Graphite:</div>
            <a href="https://app.graphite.dev/media/video/xncBfoFfDL8XmXxkoXEI/9ce42919-5b2a-4ba6-958e-b7c3eb79505e.mov">
              <img src="https://app.graphite.dev/api/v1/graphite/video/thumbnail/xncBfoFfDL8XmXxkoXEI/9ce42919-5b2a-4ba6-958e-b7c3eb79505e.mov">
            </a>
          </div>
<video src="https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xncBfoFfDL8XmXxkoXEI/9ce42919-5b2a-4ba6-958e-b7c3eb79505e.mov">Screen Recording 2024-08-12 at 19.28.50.mov</video>

## Caveats

## Test Plan