Fix problem with risk signals glossary (#11224)
Bug when opening the risk signal glossary. Reproduce: 

1. From the user nav in the dropdown select glossary, and close it with
`esc` or `close`
2. `body` tag has `style: pointer-events = none` and this prevents the
user from clicking anywhere

This happens because the drawer opens from a dropdown, which adds the
restriction of pointer events, and it begins before the dropdown
finishes closing. I solved it by adding a 50ms delay for opening.

I also update `dialog` and `dropdown` radix base component to the latest
version because there was a known problem with pointer events in the
past with these components.

Fix: 


https://github.com/user-attachments/assets/9531f070-099f-4fb8-aed0-631631d359a7