[BE-508][docs] Add ability to hide fields of API schemas behind a preview API gate (#11260)
## Describe your changes
There are some request/response fields that we don't want to display to
every tenant in the docs site. This adds a macro to the backend that
allows specifying when a field should be gated behind a tenant preview
api:
```
    #[openapi(gated = "implicit_auth")]
```

Behind the scenes, this ends up serializing information on the required
preview API inside the field's open API description:
```
[gated(implicit_auth)]: <>\nWhen true, will not require physical authentication from the user if they have logged into\n your organization recently.\n This may be useful if a user is running through multiple playbooks and has already recently\n logged in.
```

This uses a psuedo-markdown comment (`[]: <>`) to communicate that the
field is `gated(implicit_auth)`. This gating information is then
magically not rendered on the docs site. But our docs site has some code
to look for this `gated(implicit_auth)` syntax as an indicator that the
provided field is gated

## Caveats


## Test Plan
This is pretty hacky and feels easy to break. I'd love to add unit test
coverage for this, will do in follow-up PRs after we add support for
tests in the docs site