[FP-7971] (new) Business-profile uses new Form.Field logic and has tests (#11388)
## Describe your changes
See #11325 , that's this! 

---

Wholly internal change:
- Now use `Form.Field` and `Form.Label` everywhere in content for
settings page
- Clicking on labels now focuses the inputs!
- Add `horizontal` option to `Form.Field`
- `FormFieldContext` that passes id down to label and child elements (we
use to have same id between input and label)
- Lots of testing on this page! Testing that everything displays
properly, that the page is available only if folks have access, that the
mutations work (and that the succeess and failure cases work correctly),
copying works, and a11y features for clicking on the label work

## Caveats
- One thing I realized is that I need to test the logo uploading
functionality - users would expect an uploaded logo to only submit if
they click "save changes," but because the logo functions separately
from the rest of the form, that won't be the case. This will come in a
separate PR, ticketed as
[FP-7974](https://linear.app/footprint/issue/FP-7974/fix-logo-logic)
- Not sure if we love the interface of the `Form.FieldContext` stuff -
open to changes here, let me know what you think!
- Still to come: ripping out all old logic and using redirect
([FP-7965](https://linear.app/footprint/issue/FP-7965/rip-out-old-frontend-logic-for-business-profile-team-and-roles-billing)
- Do we think we need tests for the navigator? Only test would really be
seeing that the links worked - this seems plausibly overkill, but I'm
happy to add if we prefer. What do you think?

## Test Plan
- Lots of (new) testing in `content.test.tsx`
- UI!