Add vault version response header to open API documentation (#11365)
## Describe your changes
Turns the old `WithVaultVersionHeader` into a generic `ResponseWithHeaders<TBody, THeaders>`  struct that we can reuse in the future for any other responses that have headers. We keep `WithVaultVersionHeader` as an alias to `ResponseWithHeaders<TBody, ScopedVaultVersionResponseHeader>`.

This `WithVaultVersionHeader` has special logic to include the headers defined by `THeaders` in the open API response.

This also includes the `x_fp_preview_gate` indicating that the header should only be conditionally rendered on the docs site

## Caveats
In a follow-up PR, I will start to render response headers on the docs site and properly respect the preview API feature gate

## Test Plan