[docs] Pass request/response structs around in addition to schemas (#11408)
## Describe your changes
Before this change, we would unwrap an open API request/response to just its schema. When we start passing headers into the docs site, we're going to need to be able to pass the response around with its headers. So this stops unwrapping the request/response into just its schema and instead passes the whole request/response around

This doesn't change functionality, just a refactor for next PR

## Caveats

## Test Plan