[BE-555][docs] Render response headers on API reference (#11410)
## Describe your changes
Given the response headers that are now being serialized in the open API spec, start to render them in the API reference.

This also filters out response headers that are not visible due to a feature gate.

## Caveats
I did not have any design support for this... it's getting a little clunky with request headers / params / body and response body / headers. Might be time to move towards a more stripe-esque model. I will chat with claudio when he is back :)

## Test Plan
When logged in (and can see the response header):

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xncBfoFfDL8XmXxkoXEI/c49b3daf-010b-426d-9652-bd7d023958c7.png)

When logged out:

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xncBfoFfDL8XmXxkoXEI/3482bc4f-eccb-4297-b937-b96fef89c41d.png)