[FP-7995] hide billing from all tenants except for grid (#11440)
## Describe your changes

Only show billing (page, option in navigator) if user is a firm employee
or a grid admin.

## Caveats

N/A

## Test Plan

This `shouldShowBilling` is the logic we had previously. This is tricky
to test because I remain a firm employee when I test with Grid; I tried
this by removing the `isFirmEmployee` condition and seeing if it still
showed up. May need some help testing this more, though.