[FP-7986] components update react-hook-form` (#11386)
## Describe your changes

react-hook-form 7.47.0 --> 7.52.2

I had to hunt for this upstream change, and we had a real test fail!

In essence, the way that RHF handles `isDirty` in new versions is
slightly different—you can see that they distinguish between fields that
have been touched vs. those that are dirty in the changing of [this
internal
test](https://github.com/react-hook-form/react-hook-form/commit/28a45bd7fb3dbf27489cdb166a75613b8a552058)
in one of the versions between the one we previously had for components
and the one way had now.

Basically: after poking around a bit, I realized this was failing
because `isDirty` was initialized as true, even when we hadn't touched
any fields. I fixed this by using `touchedFields` in this form. I think
this may actually be broken, upstream, though (or, at least: is weird
behavior!)


## Caveats

Split into many PRs 

## Test Plan

CI