[FP-8000] Genericize field-value TextInputs (#11471)
## Describe your changes

UX no-op! This is mergeable now without other changes, just makes it
easier for us to keep working.

- Swap container to a `Form.Field` so that elements have id context
- Wrap label in a `Form.Label` so we deal with `htmlFor` case
- Use `Form.Input` instead of `TextInput` in elements that display on
`<FieldValue />`
- Genericize some of the inputs (i.e. `StateSelect`, `ZipInput`) so that
we can pass the `fieldName` down and (eventually) edit `BusinessDI`s
- `DobInput` --> `DateInput` and use for business 

## Caveats

This will not yet allow users to edit business information! 

We still have this line in `field-value`:

```
  const { value, showEditView, canEditField, isDecrypted, name: di, transforms } = field;
```

Right now, because we don't have the dropdown for "Edit user
information" for businesses—`showEditField` and `canEditField` will
never be positive for `BusinessDI`s.

This doesn't include the current logic for `NativeSelect` components
that we have in some of this form. That needs to be in a separate PR
because the `Form.Select` logic has not merged yet (see #11449)

## Test Plan

- current testing
- test in UI for a business to make sure editable version still don't
show up
- test in UI 
- pedro (✅)